Fixed socket shutdown behaviour not releasing system resources properly, socket should now properly close.
Note: .NET has not properly documented how the sockets behave so trial and error is appropriate for this suite, I will be testing this against a full test suite and hopefully providing a much more stable experience.

I've tidied up a small bit of shutdown logic in this and added LingerState handling, if a socket is in the shutdown process it can now be re-allocated to.

Since the heartbeat is sent, I believe that we should be OK to shutdown fully under 2 seconds.