Fixes which add support to handle more than 1000 connections, this degrades performance though, so I've got to add async, then thread pooling for the network sockets.
At the moment the locks are static and shared across all instances like the GIL in python.

This isn't good for performance, I'm going to do my best to remove static from the netpeer class and hopefully prevent the locks taking performance away.

Once the socket receive is async and send is async I should have a lot more headroom to play with then I can focus on pooling based on the system resources available.